+++ CHANGELOG.md
@@ -15,6 +15,11 @@
 * Inherit `StyleMix` for `Window` to facilitate theme implementation.
     - See `zng::window` documentation for theming tips.
 
+* Unify pixel resolution units.
+    - **Breaking** Removed `zng::layout::{Ppi, Dpi, ResolutionUnits}`.
+    - Added `zng::layout::{PxDensity, PxDensity2d, PxDensityUnits}`.
+    - **Breaking** Renamed all *ppi* to *density*.
+
 * Unify `MixBlendMode` type with view API.
     - **Breaking** Remove `zng::color::RenderMixBlendMode`.
     - **Breaking** `zng::color::MixBlendMode` is now non-exhaustive and has a new variant `PlusLighter`.